FaceID/requirements.txt
absl-py==1.0.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.0
bleach==5.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
comtypes==1.1.11
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dlib==19.24.0
dnspython==2.2.1
entrypoints==0.4
executing==0.8.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.15.3
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.1
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.1
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyter-server==1.16.0
jupyterlab==3.3.3
jupyterlab-pygments==0.2.0
jupyterlab-server==2.12.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mongoengine==0.24.1
mtcnn==0.1.1
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pymongo==4.1.1
pyparsing==3.0.7
pypiwin32==223
pyrsistent==0.18.1
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.1
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
tkcalendar==1.6.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
